requirements.txt
requests==2.25.1
Telethon==1.21.1
python-dotenv==0.19.0
